Prevent 'taskhash mismatch' errors during UBI creation.
The recipe writes the current DATE to a file.
This may evaluate to a different value in a subprocess, e.g. due to locale settings.
To work around that, put the date stamp into a variable at parse time and exclude it from dependency parsing.

This should solve the occasional 'taskhash mismatch' errors that occur while building.

Signed-off-by: Erik Slagter <erik@openpli.org>